Implement support for goals

+++ lib/ynab/models/budget_detail_wrapper.rb
@@ -31,7 +31,7 @@ module YNAB
     def self.swagger_types
       {
         :'budget' => :'BudgetDetail',
-        :'server_knowledge' => :'Float'
+        :'server_knowledge' => :'Integer'
       }
     end
