+++ lib/ynab/models/hybrid_transaction.rb
@@ -109,7 +109,7 @@ module YNAB
       {
         :'id' => :'String',
         :'date' => :'Date',
-        :'amount' => :'Float',
+        :'amount' => :'Integer',
         :'memo' => :'String',
         :'cleared' => :'String',
         :'approved' => :'BOOLEAN',
